+++ app/Http/Controllers/Admin/Individual/IndividualController.php
@@ -58,9 +58,9 @@ class IndividualController extends Controller
                     })
                    ->escapeColumns([])
                     ->addColumn('action', function($row){
-   
+						  $txt = "'Are you sure to delete this?'";
                           $btn = '<a href="'.url('admin').'/individual/'.$row->id.'/edit" class="edit btn btn-primary btn-sm">Edit</a>
-                                   <a href="'.url('admin').'/individual/delete/'.$row->id.'" class="delete btn btn-primary btn-sm">Delete</a>';
+                                   <a onclick="return confirm('.$txt.')" href="'.url('admin').'/individual/delete/'.$row->id.'" class="delete btn btn-primary btn-sm">Delete</a>';
      
                             return $btn;
                     })
